requirements.txt
asgiref==3.6.0
click==8.1.3
Django==3.2.10
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
pytz==2022.7.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.3
Werkzeug==2.2.2
